+++ HexaEngine.Core/Graphics/BufferExtendedShaderResourceViewFlags.cs
@@ -3,7 +3,7 @@
     [Flags]
     public enum BufferExtendedShaderResourceViewFlags : int
     {
-        Raw = unchecked((int)1),
-        None = unchecked((int)0)
+        Raw = unchecked(1),
+        None = unchecked(0)
     }
 }
